refactor: change file structure
