+++ multisig/simulations/ethereum/vip-022/abi/ResilientOracleAbi.json
