+++ multisig/simulations/sepolia/vip-022/abi/PoolRegistryAbi.json
