+++ multisig/simulations/sepolia/vip-022/abi/ResilientOracleAbi.json
